backend connection implementation
