+++ app/mobile/BuPazar/app/src/main/res/layout/fragment_cart.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".page.fragment.BasketFragment">
+    tools:context=".page.fragment.CartFragment">
 
     <RelativeLayout
         android:layout_width="match_parent"
